Created viewmodel for datasources.create
