+++ VSProjects/Geomapmaker/ViewModels/DataSources/CreateDataSourceVM.cs
@@ -1,0 +1,11 @@
+﻿using ArcGIS.Desktop.Framework.Controls;
+
+namespace Geomapmaker.ViewModels.DataSources
+{
+    public class CreateDataSourceVM : ProWindow
+    {
+        public string Source { get; set; }
+        public string Notes { get; set; }
+        public string Url { get; set; }
+    }
+}
